+++ contracts/beanstalk/storage/System.sol
@@ -166,9 +166,9 @@ struct Weather {
     uint128 lastDeltaSoil; // ───┐ 16 (16)
     uint32 lastSowTime; //       │ 4  (20)
     uint32 thisSowTime; //       │ 4  (24)
-    uint32 temp; //              │ 4  (28)
-    uint16 morningDuration; // ──┘ 2  (30/32)
-    uint256 morningControl;
+    uint64 temp; // ─────────────┘ 8  (32)
+    uint128 morningDuration; //  ───┐ 16 (16)
+    uint128 morningControl; // ─────┘ 16 (32)
     bytes32[3] _buffer;
 }
